+++ packages/terra-search-field/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Added
-  * Added Syles for invalid Search Field components.
+  * Added styles for invalid Search Field components.
 
 ## 3.97.0 - (November 13, 2023)